requirements.txt
numpy>=1.19.2
scipy>=1.5.2
matplotlib>=3.3.2
pyyaml>=6.0
